[fix] removed some files [update] refactoring pagination header handling
