+++ app/interfaces/controller/category.go
@@ -51,10 +51,7 @@ func (cc *CategoryController) Index() http.Handler {
 			cc.Logger.Error(err.Error())
 			JSONResponse(w, http.StatusInternalServerError, []byte(err.Error()))
 		}
-		w.Header().Set("Pagination-Count", pn.Count)
-		w.Header().Set("Pagination-Pagecount", pn.PageCount)
-		w.Header().Set("Pagination-Page", pn.Page)
-		w.Header().Set("Pagination-Limit", pn.Limit)
+		SetPaginationHeader(w, pn)
 		JSONResponse(w, http.StatusOK, res)
 	})
 }
@@ -76,10 +73,7 @@ func (cc *CategoryController) IndexPrivate() http.Handler {
 			cc.Logger.Error(err.Error())
 			JSONResponse(w, http.StatusInternalServerError, []byte(err.Error()))
 		}
-		w.Header().Set("Pagination-Count", pn.Count)
-		w.Header().Set("Pagination-Pagecount", pn.PageCount)
-		w.Header().Set("Pagination-Page", pn.Page)
-		w.Header().Set("Pagination-Limit", pn.Limit)
+		SetPaginationHeader(w, pn)
 		JSONResponse(w, http.StatusOK, res)
 	})
 }
