+++ app/interfaces/controller/post.go
@@ -59,10 +59,7 @@ func (pc *PostController) Index() http.Handler {
 			pc.Logger.Error(err.Error())
 			JSONResponse(w, http.StatusInternalServerError, []byte(err.Error()))
 		}
-		w.Header().Set("Pagination-Count", pn.Count)
-		w.Header().Set("Pagination-Pagecount", pn.PageCount)
-		w.Header().Set("Pagination-Page", pn.Page)
-		w.Header().Set("Pagination-Limit", pn.Limit)
+		SetPaginationHeader(w, pn)
 		JSONResponse(w, http.StatusOK, res)
 	})
 }
@@ -85,10 +82,7 @@ func (pc *PostController) IndexByKeyword() http.Handler {
 			pc.Logger.Error(err.Error())
 			JSONResponse(w, http.StatusInternalServerError, []byte(err.Error()))
 		}
-		w.Header().Set("Pagination-Count", pn.Count)
-		w.Header().Set("Pagination-Pagecount", pn.PageCount)
-		w.Header().Set("Pagination-Page", pn.Page)
-		w.Header().Set("Pagination-Limit", pn.Limit)
+		SetPaginationHeader(w, pn)
 		JSONResponse(w, http.StatusOK, res)
 	})
 }
@@ -111,10 +105,7 @@ func (pc *PostController) IndexByCategory() http.Handler {
 			pc.Logger.Error(err.Error())
 			JSONResponse(w, http.StatusInternalServerError, []byte(err.Error()))
 		}
-		w.Header().Set("Pagination-Count", pn.Count)
-		w.Header().Set("Pagination-Pagecount", pn.PageCount)
-		w.Header().Set("Pagination-Page", pn.Page)
-		w.Header().Set("Pagination-Limit", pn.Limit)
+		SetPaginationHeader(w, pn)
 		JSONResponse(w, http.StatusOK, res)
 	})
 }
@@ -137,10 +128,7 @@ func (pc *PostController) IndexByTag() http.Handler {
 			pc.Logger.Error(err.Error())
 			JSONResponse(w, http.StatusInternalServerError, []byte(err.Error()))
 		}
-		w.Header().Set("Pagination-Count", pn.Count)
-		w.Header().Set("Pagination-Pagecount", pn.PageCount)
-		w.Header().Set("Pagination-Page", pn.Page)
-		w.Header().Set("Pagination-Limit", pn.Limit)
+		SetPaginationHeader(w, pn)
 		JSONResponse(w, http.StatusOK, res)
 	})
 }
@@ -162,10 +150,7 @@ func (pc *PostController) IndexPrivate() http.Handler {
 			pc.Logger.Error(err.Error())
 			JSONResponse(w, http.StatusInternalServerError, []byte(err.Error()))
 		}
-		w.Header().Set("Pagination-Count", pn.Count)
-		w.Header().Set("Pagination-Pagecount", pn.PageCount)
-		w.Header().Set("Pagination-Page", pn.Page)
-		w.Header().Set("Pagination-Limit", pn.Limit)
+		SetPaginationHeader(w, pn)
 		JSONResponse(w, http.StatusOK, res)
 	})
 }
