+++ app/interfaces/controller/tag.go
@@ -51,10 +51,7 @@ func (tc *TagController) Index() http.Handler {
 			tc.Logger.Error(err.Error())
 			JSONResponse(w, http.StatusInternalServerError, []byte(err.Error()))
 		}
-		w.Header().Set("Pagination-Count", pn.Count)
-		w.Header().Set("Pagination-Pagecount", pn.PageCount)
-		w.Header().Set("Pagination-Page", pn.Page)
-		w.Header().Set("Pagination-Limit", pn.Limit)
+		SetPaginationHeader(w, pn)
 		JSONResponse(w, http.StatusOK, res)
 	})
 }
@@ -76,10 +73,7 @@ func (tc *TagController) IndexPrivate() http.Handler {
 			tc.Logger.Error(err.Error())
 			JSONResponse(w, http.StatusInternalServerError, []byte(err.Error()))
 		}
-		w.Header().Set("Pagination-Count", pn.Count)
-		w.Header().Set("Pagination-Pagecount", pn.PageCount)
-		w.Header().Set("Pagination-Page", pn.Page)
-		w.Header().Set("Pagination-Limit", pn.Limit)
+		SetPaginationHeader(w, pn)
 		JSONResponse(w, http.StatusOK, res)
 	})
 }
