--- app/usecase/dto/auth_request.go
@@ -1,7 +1,0 @@
-package dto
-
-// A RequestCredential represents the singular of credential.
-type RequestCredential struct {
-	Email    string `json:"email"`
-	Password string `json:"password"`
-}
