--- app/usecase/dto/category_request.go
@@ -1,6 +1,0 @@
-package dto
-
-// A RequestCategory represents the singular of post for store.
-type RequestCategory struct {
-	Name string `json:"name"`
-}
