--- app/usecase/dto/http_status_response.go
@@ -1,6 +1,0 @@
-package dto
-
-// A ResponseHTTPStatus represents the singular of http status response.
-type ResponseHTTPStatus struct {
-	Message string `json:"message"`
-}
