--- app/usecase/dto/json_response.go
@@ -1,8 +1,0 @@
-package dto
-
-import "net/http"
-
-// A JSONResponse is a presenter interface for response.
-type JSONResponse interface {
-	HTTPStatus(w http.ResponseWriter, code int, msg []byte)
-}
