--- app/usecase/dto/tag_request.go
@@ -1,6 +1,0 @@
-package dto
-
-// A RequestTag represents the singular of tag for store.
-type RequestTag struct {
-	Name string `json:"name"`
-}
